requirements.txt
#usage: pip install -r requirements.txt

alembic == 1.1.0
bcrypt == 3.1.7
cffi == 1.12.3
Click == 7.0
coverage == 4.5.4
Flask == 1.1.1
Flask-Bcrypt == 0.7.1
Flask-Migrate == 2.5.2
Flask-Script == 2.0.6
Flask-SQLAlchemy == 2.4.0
Flask-Testing == 0.7.1
Flask-Cors == 3.0.8
gunicorn == 19.9.0
itsdangerous == 1.1.0
Jinja2 == 2.11.3
Mako == 1.1.0
MarkupSafe == 1.1.1
nose == 1.3.7
pdfkit == 0.6.1
Pillow == 8.1.1
pycparser == 2.19
pyfakefs == 3.6
PyJWT == 1.7.1
python-dateutil == 2.8.0
python-editor == 1.0.4
six == 1.12.0
SQLAlchemy == 1.3.8
Werkzeug == 0.15.6
